pyproject.toml
[project]
name = "stealth-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "playwright>=1.42.0",
    "psutil>=5.9.8",
    "pyyaml>=6.0.1",
    "redis>=5.0.3",
    "seleniumbase>=4.28.0",
    "nest_asyncio>=1.6.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pytest>=8.0.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "httpx>=0.28.1",
]
